tools/embedder/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==24.3.0
beautifulsoup4==4.12.3
boto3==1.35.98
botocore==1.35.98
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
filelock==3.16.1
fsspec==2024.12.0
gpt4all==2.8.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
keybert==0.8.5
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langsmith==0.2.10
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
nomic==3.3.4
numpy==2.2.1
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pgvector==0.3.6
pillow==11.1.0
psycopg2==2.9.10
pyarrow==18.1.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.1
pymupdf4llm==0.0.17
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
strip-markdown==1.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
timescale-vector==0.0.7
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
